# laravel/task_Agiletech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/district": {
            "get": {
                "tags": [
                    "district"
                ],
                "summary": "Get a list of districts",
                "operationId": "09b3bc4dc23bef12c24ffffbfa09aee2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/district/": {
            "post": {
                "summary": "Store a newl district created resource in storage.",
                "operationId": "42e615ec18127e6ff8333764644eb369",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "district's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "district's region_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The district was successfully created"
                    }
                }
            }
        },
        "/api/district/id": {
            "get": {
                "tags": [
                    "district"
                ],
                "summary": "Display the details of district.",
                "operationId": "3c5bdea9c0479d94e5696b9e365e59a9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  district in storage.",
                "operationId": "e4cb5618870051b861292bf6bb7681b9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "district's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "district's region_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The district was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "district"
                ],
                "summary": "Remove the district from storage.",
                "operationId": "8d125c5c8609decf8ff4b9347f6e16ac",
                "responses": {
                    "200": {
                        "description": "The district was successfully deleted"
                    }
                }
            }
        },
        "/api/img": {
            "get": {
                "tags": [
                    "img"
                ],
                "summary": "Get a list of imgs",
                "operationId": "77b12b0b1f285163ff4a880afd272fa0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/img/": {
            "post": {
                "summary": "Store a newl img created resource in storage.",
                "operationId": "e79faef017620a693e9810b48a21412c",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "img's branch_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img",
                        "in": "query",
                        "description": "img's img",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The img was successfully created"
                    }
                }
            }
        },
        "/api/img/id": {
            "get": {
                "tags": [
                    "img"
                ],
                "summary": "Display the details of img.",
                "operationId": "c1e9912fac9c0fc38bd5e6fe5810f1ec",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  img in storage.",
                "operationId": "abcfe07c85b26aa9b67ed65cc8deaa7f",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "img's branch_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img",
                        "in": "query",
                        "description": "img's imgs",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The img was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "img"
                ],
                "summary": "Remove the img from storage.",
                "operationId": "ac407ca5e1e7150dd19dcee3dbcb2260",
                "responses": {
                    "200": {
                        "description": "The img was successfully deleted"
                    }
                }
            }
        },
...
        "/api/region": {
            "get": {
                "tags": [
                    "region"
                ],
                "summary": "Get a list of regions",
                "operationId": "0c51f58df055e5721d4627d15b7ceb04",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/regiob/": {
            "post": {
                "summary": "Store a newl regiob created resource in storage.",
                "operationId": "526a359bde5298e0459c984a35b28916",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "regiob's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The regiob was successfully created"
                    }
                }
            }
        },
        "/api/region/id": {
            "get": {
                "tags": [
                    "region"
                ],
                "summary": "Display the details of region.",
                "operationId": "619f156bb5b5ba524071abca19110001",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  region in storage.",
                "operationId": "80ec33d5990143acc98883b1853a03e5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "region's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "region's region_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The region was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "region"
                ],
                "summary": "Remove the region from storage.",
                "operationId": "85d145c3a524b8f1bdc2155973660b75",
                "responses": {
                    "200": {
                        "description": "The region was successfully deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}